requirements.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.10
attrs==20.2.0
audioread==2.1.9
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cycler==0.10.0
decorator==4.4.2
Django==3.1.2
django-autocomplete-light==3.5.1
django-crispy-forms==1.9.2
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
djangorestframework-gis==0.16
hypothesis==5.37.3
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
irekua-database==0.1.2
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.34.0
MarkupSafe==1.1.1
matplotlib==3.3.2
numba==0.51.2
numpy==1.19.2
packaging==20.4
pandas==1.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
pooch==1.2.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycountry==20.7.3
pycparser==2.20
Pygments==2.7.1
pymediainfo==4.2.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
resampy==0.2.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sorl-thumbnail==12.6.3
sortedcontainers==2.2.2
SoundFile==0.10.3.post1
sqlparse==0.4.1
threadpoolctl==2.1.0
timezonefinder==4.4.1
tqdm==4.51.0
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.2.5
